Move ErrorUtilities_Tests to Assumed_Tests
ErrorUtilities_Tests.cs tested Assumed.True APIs, not ErrorUtilities.
Consolidate these tests into Assumed_Tests.cs in Framework.UnitTests,
updating from xUnit Assert to Shouldly and adding coverage for:
- True(bool, string?) with a non-null message (both passing and failing)
- Lazy evaluation of the TrueInterpolatedStringHandler: verifies the
  interpolated string is not formatted when the condition is true, is
  formatted when false, and formats multiple args correctly

Remove ErrorUtilities_Tests.cs and its <Compile> links from
Build.UnitTests, Tasks.UnitTests, and Utilities.UnitTests.
