+++ src/Tasks.UnitTests/Microsoft.Build.Tasks.UnitTests.csproj
@@ -45,7 +45,6 @@
 
   <ItemGroup>
     <Compile Include="..\Shared\UnitTests\AssemblyNameEx_Tests.cs" />
-    <Compile Include="..\Shared\UnitTests\ErrorUtilities_Tests.cs" />
     <Compile Include="..\Shared\UnitTests\PrintLineDebugger_Tests.cs" />
     <Compile Include="..\Shared\UnitTests\NativeMethodsShared_Tests.cs">
       <Link>NativeMethodsShared_Tests.cs</Link>
